+++ mpp/src/CircleBatch.cpp
@@ -168,7 +168,7 @@ namespace mpp
 				size);
 		}
 
-		for (int i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
+		for (size_t i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
 		{
 			createVertexBuffer(i, mesh, vertexCount, false);
 		}
