+++ mpp/src/QuadBatch.cpp
@@ -372,7 +372,7 @@ namespace mpp
 				mPointSize);
 		}
 
-		for (int i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
+		for (size_t i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
 		{
 			auto const& layout = mSpecification.getVertexBufferAttributeLayout(i);
 			createVertexBuffer(i, mesh, vertexCount, layout.isStatic());
